+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - granular JWT scopes for services: cov-survey, program-management, user-preferences, consent-management, cds-userdata
+- IP address obfuscator
 
 ### Changed